+++ examples/go/clients/ton/main.go
@@ -8,9 +8,9 @@ import (
 	"os"
 	"time"
 
-	t402 "github.com/awesome-doge/t402/go"
-	"github.com/awesome-doge/t402/go/mechanisms/ton"
-	tonclient "github.com/awesome-doge/t402/go/mechanisms/ton/exact/client"
+	t402 "github.com/t402-io/t402/go"
+	"github.com/t402-io/t402/go/mechanisms/ton"
+	tonclient "github.com/t402-io/t402/go/mechanisms/ton/exact/client"
 	"github.com/joho/godotenv"
 )