+++ go/http/gin/middleware.go
@@ -8,9 +8,9 @@ import (
 	"sync"
 	"time"
 
-	t402 "github.com/awesome-doge/t402/go"
-	"github.com/awesome-doge/t402/go/extensions/bazaar"
-	t402http "github.com/awesome-doge/t402/go/http"
+	t402 "github.com/t402-io/t402/go"
+	"github.com/t402-io/t402/go/extensions/bazaar"
+	t402http "github.com/t402-io/t402/go/http"
 	"github.com/gin-gonic/gin"
 )