+++ go/http/server.go
@@ -11,8 +11,8 @@ import (
 	"strconv"
 	"strings"
 
-	t402 "github.com/awesome-doge/t402/go"
-	"github.com/awesome-doge/t402/go/types"
+	t402 "github.com/t402-io/t402/go"
+	"github.com/t402-io/t402/go/types"
 )
 
 // ============================================================================