+++ go/mechanisms/evm/exact/server/scheme.go
@@ -7,9 +7,9 @@ import (
 	"strconv"
 	"strings"
 
-	t402 "github.com/awesome-doge/t402/go"
-	"github.com/awesome-doge/t402/go/mechanisms/evm"
-	"github.com/awesome-doge/t402/go/types"
+	t402 "github.com/t402-io/t402/go"
+	"github.com/t402-io/t402/go/mechanisms/evm"
+	"github.com/t402-io/t402/go/types"
 )
 
 // ExactEvmScheme implements the SchemeNetworkServer interface for EVM exact payments (V2)