+++ go/mechanisms/ton/exact/facilitator/scheme.go
@@ -7,9 +7,9 @@ import (
 	"strconv"
 	"time"
 
-	t402 "github.com/awesome-doge/t402/go"
-	"github.com/awesome-doge/t402/go/mechanisms/ton"
-	"github.com/awesome-doge/t402/go/types"
+	t402 "github.com/t402-io/t402/go"
+	"github.com/t402-io/t402/go/mechanisms/ton"
+	"github.com/t402-io/t402/go/types"
 )
 
 // ExactTonScheme implements the SchemeNetworkFacilitator interface for TON exact payments (V2)